% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbiom.R
\name{loadFromBiom}
\alias{loadFromBiom}
\title{load biom file into a \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}}
\usage{
loadFromBiom(file)
}
\arguments{
\item{file}{biom file location}
}
\value{
An object of class \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}
}
\description{
load biom file into a \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}
}
\examples{
if(requireNameSpace("biomformat")) {
  rich_dense_file  = system.file("extdata", "rich_dense_otu_table.biom", package = "biomformat")
  loadFromBiom(rich_dense_file)
}
}
