% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbiom.R
\name{makeMicrobiomeExperimentFromBiom}
\alias{makeMicrobiomeExperimentFromBiom}
\title{coerce \code{\link[biomformat]{biom}} into a \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}}
\usage{
makeMicrobiomeExperimentFromBiom(obj)
}
\arguments{
\item{obj}{object of type \code{\link[biomformat]{biom}}}
}
\value{
An object of class \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}
}
\description{
coerce \code{\link[biomformat]{biom}} into a \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}
}
\examples{
if(requireNameSpace("biomformat")) {
  rich_dense_file  = system.file("extdata", "rich_dense_otu_table.biom", package = "biomformat")
  x1 <- read_biom(rich_dense_file)
  as(x1, "MicrobiomeExperiment")
}
}
