% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMicrobiomeExperimentFromphyloseq.R
\name{makeMicrobiomeExperimentFromphyloseq}
\alias{makeMicrobiomeExperimentFromphyloseq}
\title{Coerce phyloseq object}
\usage{
makeMicrobiomeExperimentFromphyloseq(object)
}
\arguments{
\item{object}{}
}
\value{
An object of class MicrobiomeExperiment
}
\description{
Coerce phyloseq object
}
\examples{
if (requireNamespace("phyloseq")) {
    data(GlobalPatterns, package="phyloseq")
    as(GlobalPatterns, "MicrobiomeExperiment")
    data(enterotype, package="phyloseq")
    as(enterotype, "MicrobiomeExperiment")
    data(esophagus, package="phyloseq")
    as(esophagus, "MicrobiomeExperiment")
}
}
