% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-methods.R
\docType{methods}
\name{MEcounts}
\alias{MEcounts}
\title{Counts accessor for MicrobiomeExperiment class}
\usage{
\S4method{counts}{MicrobiomeExperiment}(object, ...)
}
\arguments{
\item{x}{an object of class MicrobiomeExperiment}
}
\description{
~description~
}
