% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment.R
\docType{package}
\name{MicrobiomeExperiment-package}
\alias{MicrobiomeExperiment-package}
\title{MicrobiomeExperiment Package.}
\description{
Package defining MicrobiomeExperiment class and methods
}
\examples{
print("We have a package!")
}
\seealso{
\link[MicrobiomeExperiment]{MicrobiomeExperiment-class}
}
\author{
Nathan D. Olson \email{nolson@umiacs.umd.edu}
}
\keyword{package}
