% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-class.R
\name{MicrobiomeExperiment}
\alias{MicrobiomeExperiment}
\alias{MicrobiomeExperiment-class}
\title{The MicrobiomeExperiment representation class}
\usage{
MicrobiomeExperiment(assays = SimpleList(), rowData = MicrobiomeFeatures(),
  colData = DataFrame(), metadata = list())
}
\description{
SummarizedExperiment-like class for microbiome data. rowData is
a MicrobiomeFeatures object LINK so it includes: a taxonomy table
(DataFrame), #' optional phylogentic tree (phylo object), and a sequence
database.
}
\details{
It supports (most) of the interface to phyloseq objects
}
\examples{

library(metagenomeFeatures)
data(mock_mgF)

sampleNames <- letters[1:4]
pd <- DataFrame(a=letters[1:4], b=1:4)
numcounts <- nrow(mock_mgF) * 4
counts <- matrix(sample(1:1000,numcounts,replace=TRUE), nr=nrow(mock_mgF), nc=4)

MicrobiomeExperiment(assays=SimpleList(counts=counts),
rowData=mock_mgF,
colData=pd
)

}
