% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeFeatures-class.R
\name{MicrobiomeFeatures-class}
\alias{MicrobiomeFeatures-class}
\alias{MicrobiomeFeatures}
\title{The MicrobiomeFeatures class}
\usage{
MicrobiomeFeatures(taxa = DataFrame(), tree = NULL, seq = NULL)
}
\arguments{
\item{taxa}{taxonomic information as a \code{DataFrame}.
See \code{\link[metagenomeFeatures:mgFeatures-class]{mgFeatures}} for more
details.}

\item{tree}{phylogenetic tree. See
\code{\link[metagenomeFeatures:mgFeatures-class]{mgFeatures}} for more
details.}

\item{seq}{reference sequence. See
\code{\link[metagenomeFeatures:mgFeatures-class]{mgFeatures}} for more}
}
\description{
Alias for mgFeatures class from metagenomeFeatures package
}
