% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeFeatures-class.R
\name{MicrobiomeFeatures}
\alias{MicrobiomeFeatures}
\alias{MicrobiomeFeatures-class}
\title{Alias for mgFeatures class from metagenomeFeatures package}
\usage{
MicrobiomeFeatures(taxa = DataFrame(), phyloTree = NULL, seq = NULL,
  metadata = list())
}
\description{
Alias for mgFeatures class from metagenomeFeatures package
}
