% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleConcentrations.R
\name{sampleConcentrations}
\alias{sampleConcentrations}
\alias{sampleConcentrations<-}
\alias{sampleConcentrations,MicrobiomeExperiment-method}
\alias{sampleConcentrations<-,MicrobiomeExperiment-method}
\title{Sample concentration methods}
\usage{
sampleConcentrations(object, ...)

sampleConcentrations(object, ...) <- value

\S4method{sampleConcentrations}{MicrobiomeExperiment}(object)

\S4method{sampleConcentrations}{MicrobiomeExperiment}(object, ...) <- value
}
\arguments{
\item{object}{A \code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}}
object.}

\item{...}{Additional arguments, currently not used.}

\item{value}{A numeric vector of length equal to \code{ncol(object)},
containing sample concentrations.}
}
\value{
For \code{sampleConcentrations} a numeric vector or \code{NULL}, if no
information are set.

For \code{sampleConcentrations<-}, a modified \code{object} is returned with
sample concentrations stored as a column in \code{\link{colData}}.
}
\description{
Getter and setter for information on sample concentrations in a
\code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}} object.
}
\details{
ToDO
}
\examples{
data(soilrep)
soilrep <- as(soilrep,"MicrobiomeExperiment")
sampleConcentrations(soilrep) <- runif(ncol(soilrep))
sampleConcentrations(soilrep)
}
\seealso{
\code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}}
}
\author{
Felix G.M. Ernst
}
