% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy-methods}
\alias{taxonomy-methods}
\alias{taxonomyRanks,SummarizedExperiment-method}
\alias{taxonomicRanks}
\alias{taxonomyRankEmpty,TreeSummarizedExperiment-method}
\alias{taxonomyRankEmpty}
\title{Taxonomy related functions}
\usage{
\S4method{taxonomyRanks}{SummarizedExperiment}(x)

\S4method{taxonomyRankEmpty}{TreeSummarizedExperiment}(
  x,
  rank = taxonomyRanks(x)[1],
  empty.fields = c(NA, "", " ", "\\t", "-")
)
}
\arguments{
\item{x}{\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}
}
\value{
\itemize{
  \item{\code{taxonomyRanks}:} {a \code{character} vector with all the
    taxonomic ranks found in \code{colnames(rowData(x))}}
  \item{\code{taxonomyRankEmpty}:} {a \code{logical} value}
}
}
\description{
These function work on optional data present in \code{rowData}.

\code{taxonomyRanks} returns, which columns of \code{rowData(x)} are regarded
as columns containing taxonomic information.

\code{taxonomyRankEmpty} checks, if a selected rank is empty of information.
}
\examples{
data(esophagus)
esophagus
plot(rowTree(esophagus))
# get a factor for merging
f <- factor(regmatches(rownames(esophagus),
                       regexpr("^[0-9]*_[0-9]*",rownames(esophagus))))
merged <- mergeRows(esophagus,f)
plot(rowTree(merged))
#
data(GlobalPatterns)
GlobalPatterns
merged <- mergeCols(GlobalPatterns,colData(GlobalPatterns)$SampleType)
merged
}
\seealso{
\code{\link[=agglomerate-methods]{agglomerateByRank}}
}
