% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMicrobiomeExperimentFromMRexperiment.R
\name{makeMicrobiomeExperimentFromMRexperiment}
\alias{makeMicrobiomeExperimentFromMRexperiment}
\title{Coerce from \code{\link[metagenomeSeq]{MRexperiment}} to \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}}
\usage{
makeMicrobiomeExperimentFromMRexperiment(obj)
}
\arguments{
\item{obj}{object of type \code{\link[metagenomeSeq]{MRexperiment}}}
}
\value{
An object of class \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}
}
\description{
Coerce from \code{\link[metagenomeSeq]{MRexperiment}} to \code{\link[MicrobiomeExperiment]{MicrobiomeExperiment}}
}
\examples{
if(requireNameSpace("metagenomeSeq")) {
  data("mouseData")
  as(mouseData, "MicrobiomeExperiment")
}
}
