% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge-methods}
\alias{merge-methods}
\alias{mergeRows,SummarizedExperiment-method}
\alias{mergeCols,SummarizedExperiment-method}
\title{Merge a subset of the species in \code{x} into one species/taxa/OTU.}
\usage{
\S4method{mergeRows}{SummarizedExperiment}(x, f, archetype = 1L, mergeTree = FALSE)

\S4method{mergeCols}{SummarizedExperiment}(x, f, archetype = 1L, mergeTree = FALSE)
}
\arguments{
\item{x}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}}

\item{f}{A factor for merging. Must be the same length as
\code{nrow(x)/ncol(x)}. Row corresponding to the same level will be merged.
If \code{length(levels(f)) == nrow(x)/ncol(x)}, \code{x} will be returned
unchanged.}

\item{archetype}{Of each level of \code{f}, which element should be regarded
as the archetype and metadata in the columns or rows kept, while merging?
This can be single interger value or an integer vector of the same length
as \code{levels(f)}. (Default: \code{archetype = 1L}, which means the first
element encountered will be kept)}

\item{mergeTree}{\code{TRUE} or \code{FALSE}: should to
\code{rowTree()} also be merged? (Default: \code{mergeTree = FALSE})}
}
\value{
an object with the same class \code{x} with the specified entries
  merged into one entry in all relevant components.
}
\description{
In the context of microbiome analysis, it might be desirable to merge data
based on taxonomic levels. \code{mergeRows} uses
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
and sub-classes, which contains certain field in its \code{rowData} (See
\code{\link[=taxonomy-methods]{taxonomyRanks}} for more details).
}
\examples{
data(esophagus)
esophagus
plot(rowTree(esophagus))
# get a factor for merging
f <- factor(regmatches(rownames(esophagus),
                       regexpr("^[0-9]*_[0-9]*",rownames(esophagus))))
merged <- mergeRows(esophagus,f)
plot(rowTree(merged))
#
data(GlobalPatterns)
GlobalPatterns
merged <- mergeCols(GlobalPatterns,colData(GlobalPatterns)$SampleType)
merged
}
\seealso{
\code{\link[=agglomerate-methods]{agglomerateByRank}}
}
