% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-class.R
\name{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment}
\alias{microbiomeData}
\alias{microbiomeData,MicrobiomeExperiment-method}
\alias{microbiomeData<-}
\alias{microbiomeData<-,MicrobiomeExperiment-method}
\title{The MicrobiomeExperiment class}
\usage{
MicrobiomeExperiment(..., microbiomeData = list())

microbiomeData(x)

\S4method{microbiomeData}{MicrobiomeExperiment}(x)

microbiomeData(x) <- value

\S4method{microbiomeData}{MicrobiomeExperiment}(x) <- value
}
\arguments{
\item{...}{Arguments passed to \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}

\item{microbiomeData}{a \code{MicrobiomeFeatures} object or some object
coercible to a \code{MicrobiomeFeatures} object. See
\code{\link[=MicrobiomeFeatures-class]{MicrobiomeFeatures}} for more
details.}

\item{x}{a \code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}}
object}

\item{value}{a a \code{\link[=MicrobiomeFeatures-class]{MicrobiomeFeatures}}
object or an object coercible to one.}
}
\description{
SummarizedExperiment-like class for microbiome data.

Methods to get or set microbiome data on a
\code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}} object.
}
\examples{
library(metagenomeFeatures)
data(mock_mgF)

sampleNames <- letters[1:4]
pd <- DataFrame(a=letters[1:4], b=1:4)
numcounts <- nrow(mock_mgF) * 4
counts <- matrix(sample(1:1000, numcounts, replace=TRUE),
                 nr = nrow(mock_mgF), nc = 4)

MicrobiomeExperiment(assays = SimpleList(counts = counts),
                     rowData = mock_mgF,
                     colData = pd )
}
