% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-class.R
\name{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment}
\alias{microbiomeData}
\alias{microbiomeData,MicrobiomeExperiment-method}
\alias{microbiomeData<-}
\alias{microbiomeData<-,MicrobiomeExperiment-method}
\title{The MicrobiomeExperiment class}
\usage{
MicrobiomeExperiment(..., microbiomeData = list())

microbiomeData(x)

\S4method{microbiomeData}{MicrobiomeExperiment}(x)

microbiomeData(x) <- value

\S4method{microbiomeData}{MicrobiomeExperiment}(x) <- value
}
\description{
SummarizedExperiment-like class for microbiome data. microbiomeData is
a MicrobiomeFeatures object LINK so it includes: a taxonomy table
(DataFrame), #' optional phylogentic tree (phylo object), and a sequence
database.
}
\details{
It supports (most) of the interface to phyloseq objects
}
\examples{
library(metagenomeFeatures)
data(mock_mgF)

sampleNames <- letters[1:4]
pd <- DataFrame(a=letters[1:4], b=1:4)
numcounts <- nrow(mock_mgF) * 4
counts <- matrix(sample(1:1000, numcounts, replace=TRUE),
                 nr = nrow(mock_mgF), nc = 4)

MicrobiomeExperiment(assays = SimpleList(counts = counts),
                     rowData = mock_mgF,
                     colData = pd )
}
