% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-class.R
\docType{class}
\name{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment}
\title{The \code{MicrobiomeExperiment} class}
\usage{
MicrobiomeExperiment(..., referenceSeq = NULL)
}
\arguments{
\item{...}{Arguments passed to
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}

\item{referenceSeq}{a \code{DNAStringSet} object or some object
coercible to a \code{DNAStringSet} object. See
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}} for more
details.}
}
\description{
The \code{MicrobiomeExperiment} class is a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}-like
class for microbiome data inheriting from
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
and
\code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}}.
}
\details{
\code{MicrobiomeExperiment} adds to the above mentioned class a slot for
storing a reference sequence as a
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}.
}
\examples{
data(taxa)
sampleNames <- letters[1:4]
pd <- DataFrame(a=letters[1:4], b=1:4)
numcounts <- nrow(taxa) * 4
counts <- matrix(sample(1:1000, numcounts, replace=TRUE),
                 nr = nrow(taxa), nc = 4)
refSeq <- DNAStringSetList(one = DNAStringSet(c("A","A","A","A","A")),
                           two = DNAStringSet(c("A","A","A","A","A")))
me <- MicrobiomeExperiment(assays = SimpleList(counts = counts),
                           rowData = taxa,
                           colData = pd,
                           referenceSeq = refSeq)
me
}
