% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-class.R
\docType{class}
\name{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment}
\title{The MicrobiomeExperiment class}
\usage{
MicrobiomeExperiment(..., referenceSeq = NULL)
}
\arguments{
\item{...}{Arguments passed to \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}

\item{referenceSeq}{a \code{DNAStringSet} object or some object
coercible to a \code{DNAStringSet} object. See
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}} for more
details.}
}
\description{
SummarizedExperiment-like class for microbiome data.
}
\examples{
data(taxa)

sampleNames <- letters[1:4]
pd <- DataFrame(a=letters[1:4], b=1:4)
numcounts <- nrow(taxa) * 4
counts <- matrix(sample(1:1000, numcounts, replace=TRUE),
                 nr = nrow(taxa), nc = 4)

MicrobiomeExperiment(assays = SimpleList(counts = counts),
                     rowData = taxa,
                     colData = pd)
}
