% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment-class}
\title{MicrobiomeExperiment-class}
\description{
MicrobiomeExperiment-class
}
\examples{

library(metagenomeFeatures)
data(mock_mgF)

sampleNames <- letters[1:4]
pd <- DataFrame(a=letters[1:4], b=1:4)
numcounts <- nrow(mock_mgF) * 4
counts <- matrix(sample(1:1000,numcounts,replace=TRUE), nr=nrow(mock_mgF), nc=4)

MicrobiomeExperiment(assays=SimpleList(counts=counts),
rowData=mock_mgF,
colData=pd
)

}
