% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment.R
\docType{package}
\name{MicrobiomeExperiment-package}
\alias{MicrobiomeExperiment-package}
\title{MicrobiomeExperiment Package.}
\description{
Wrapper package for Microbiome analysis
}
\seealso{
\link[TreeSummarizedExperiment:MicrobiomeExperiment-class]{MicrobiomeExperiment} class
}
