% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabundance.R
\name{relabundance}
\alias{relabundance}
\alias{relabundance<-}
\alias{relAbundanceCounts}
\alias{relabundance,TreeSummarizedExperiment-method}
\alias{relabundance<-,TreeSummarizedExperiment-method}
\alias{relAbundanceCounts,TreeSummarizedExperiment-method}
\title{Getter / setter for relative abundance data}
\usage{
relabundance(x, ...)

relabundance(x) <- value

relAbundanceCounts(x, abund_values = "counts", name)

\S4method{relabundance}{TreeSummarizedExperiment}(x)

\S4method{relabundance}{TreeSummarizedExperiment}(x) <- value

\S4method{relAbundanceCounts}{TreeSummarizedExperiment}(x, abund_values = "counts", name)
}
\description{
\code{relabundance} is a getter/setter for relative abundance stored in the
assay slot \sQuote{relabundance} of a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object
}
\examples{
data(GlobalPatterns)
GlobalPatterns <- relAbundanceCounts(GlobalPatterns)
relabundance(GlobalPatterns)
}
