% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy-methods}
\alias{taxonomy-methods}
\alias{taxonomyRanks}
\alias{taxonomyRanks,SummarizedExperiment-method}
\alias{taxonomicRanks}
\alias{taxonomyRankEmpty}
\alias{taxonomyRankEmpty,SummarizedExperiment-method}
\alias{checkTaxonomy}
\alias{checkTaxonomy,SummarizedExperiment-method}
\title{Taxonomy related functions}
\usage{
taxonomyRanks(x)

\S4method{taxonomyRanks}{SummarizedExperiment}(x)

taxonomyRankEmpty(
  x,
  rank = taxonomyRanks(x)[1L],
  empty.fields = c(NA, "", " ", "\\t", "-")
)

\S4method{taxonomyRankEmpty}{SummarizedExperiment}(
  x,
  rank = taxonomyRanks(x)[1],
  empty.fields = c(NA, "", " ", "\\t", "-")
)

checkTaxonomy(x, ...)

\S4method{checkTaxonomy}{SummarizedExperiment}(x, checkWrap = TRUE)
}
\arguments{
\item{x}{\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}

\item{rank}{a single character defining a taxonomic rank. Must be a value of
\code{taxonomicRanks()} function.}

\item{empty.fields}{a \code{character} value defining, which values should be
regarded as empty. (Default: \code{c(NA, "", " ", "\t")}). They will be
removed if \code{na.rm = TRUE} before agglomeration.}

\item{checkWrap}{\code{TRUE} or \code{FALSE}: Should the wrapping characters
be removed, before checking? (default: \code{checkWrap = TRUE})}
}
\value{
\itemize{
  \item{\code{taxonomyRanks}:} {a \code{character} vector with all the
    taxonomic ranks found in \code{colnames(rowData(x))}}
  \item{\code{taxonomyRankEmpty}:} {a \code{logical} value}
}
}
\description{
These function work on optional data present in \code{rowData}.

\code{taxonomyRanks} returns, which columns of \code{rowData(x)} are regarded
as columns containing taxonomic information.

\code{taxonomyRankEmpty} checks, if a selected rank is empty of information.
}
\examples{
data(esophagus)
esophagus
plot(rowTree(esophagus))
# get a factor for merging
f <- factor(regmatches(rownames(esophagus),
                       regexpr("^[0-9]*_[0-9]*",rownames(esophagus))))
merged <- mergeRows(esophagus,f)
plot(rowTree(merged))
#
data(GlobalPatterns)
GlobalPatterns
merged <- mergeCols(GlobalPatterns,colData(GlobalPatterns)$SampleType)
merged
}
\seealso{
\code{\link[=agglomerate-methods]{agglomerateByRank}}
}
